+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Allow --no-config to be removed
+
 # v0.0.157 - 2019-10-01
 - Further improvements to regex chars in tags